+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## In development
 
+- Modified lessondefs api to output all lessons in one call - [#18](https://github.com/nre-learning/syringe/pull/18)
+
 ### Curriculum
 
 ### Other