+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## In development
 
 - Modified lessondefs api to output all lessons in one call - [#18](https://github.com/nre-learning/syringe/pull/18)
+- Make iframe resource image configurable - [#19](https://github.com/nre-learning/syringe/pull/19)
 
 ### Curriculum