+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Modified lessondefs api to output all lessons in one call - [#18](https://github.com/nre-learning/syringe/pull/18)
 - Make iframe resource image configurable - [#19](https://github.com/nre-learning/syringe/pull/19)
+- Use ingresses for all iframe resources - [#28](https://github.com/nre-learning/syringe/pull/28)
 
 ### Curriculum